feat: Move contextlines code to sidecar
