+++ packages/overlay/package.json
@@ -32,6 +32,7 @@
     "@sentry/types": "^8.0.0-alpha.7",
     "@sentry/utils": "^8.0.0-alpha.7",
     "@spotlightjs/tsconfig": "workspace:*",
+    "@spotlightjs/sidecar": "workspace:*",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
