+++ packages/overlay/src/constants.ts
@@ -5,5 +5,3 @@ export const DEFAULT_EXPERIMENTS = {
 };
 
 export const DEFAULT_ANCHOR = 'bottomRight';
-
-export const CONTEXT_LINES_ENDPOINT = '/@spotlight/contextlines';
